Update project version to newest Unity version. Moved all helper classes to individual files so they are easier to maintain.
Reorganized state and transition class for inspector organization and hide values that should not be changed in the inspector.
Can now select and remove transitions.
Transitions changer color when selected.
Inspectors now show if the state machine, state, or a transition is selected.
State names now appear at the center of the state node instead of the top left corner.